% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHC_dissim.R
\name{AHC_dissim}
\alias{AHC_dissim}
\title{Classification by dissimilarities matrix}
\usage{
AHC_dissim(D = 0, X = 0, method = "average")
}
\arguments{
\item{D}{dissimilarities matrix if already computed}

\item{X}{Data matrix (only if the similarity matrix S is not given)}

\item{method}{method for calculating dissimilarities between clusters (e.g. "single", "complete" ..), default = "average"}
}
\value{
list containing partition matrix and vector of cophenetic distances
}
\description{
Performs a ascending hierarchical classification by classical algorithm from the dissimilarities matrix
}
\examples{
X = as.matrix(iris[-5])
method = "single"
P = AHC_dissim(X=X, method = "median")

}
