% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHC_sim.R
\name{AHC_sim}
\alias{AHC_sim}
\title{Classification by similarities}
\usage{
AHC_sim(
  S = 0,
  X = 0,
  sparsity = 0,
  method = "average",
  Kernel.FUN = "linear",
  ...
)
}
\arguments{
\item{S}{Similarity matrix if already computed}

\item{X}{Data matrix (only if the similarity matrix S is not given)}

\item{sparsity}{threshold below which similarities are zero}

\item{method}{method for calculating dissimilarities between clusters (e.g. "single", "complete" ..), default = "average"}

\item{Kernel.FUN}{kernel function to be used if we want to calculate S from X (default :"linear")}

\item{...}{other parameters that will be passed to the Kernel procedure (see Kernel procedure for details)}
}
\value{
P : list containing parition matrix and vector of kophenetic distances
}
\description{
Performs a ascending hierarchical classification by the classical algorithm from the similarity matrix
}
\examples{
X = as.matrix(iris[-5])
method = "single"
P = AHC_sim(X=X, method = "median", Kernel.FUN = "sigmoid")

}
