% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophcor.R
\name{cophcor}
\alias{cophcor}
\title{Cophenetic correlation}
\usage{
cophcor(hclust1, hclust2)
}
\arguments{
\item{hclust1}{first list having as elements a matrix P of partitions and a vector coph of kinetic distances}

\item{hclust2}{second list having as elements a matrix P of partitions and a vector coph of kinetic distances}
}
\value{
cophenetic correlation between the dendrograms of hclust1 and hclust2
}
\description{
Computes the cophenetic correlation between two dendograms
}
