% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dissim}
\alias{dissim}
\title{Dissimilarity method using C++}
\usage{
dissim(X, method = "average", dissim_matrix = FALSE)
}
\arguments{
\item{X}{Data matrix by default (if dissim_matrix = True, then this is the dissimilarity matrix)}

\item{dissim_matrix}{false by default, put true if you give the dissimilarity matrix directly}

\item{String}{method type of linkage for the Lance-Williams distance ("average" by default)}
}
\value{
P clustering matrix (each row corresponds to each iteration)
}
\description{
Dissimilarity method using C++
}
