% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.to.tree.R
\name{hclust.to.tree}
\alias{hclust.to.tree}
\title{Classification by hclust package}
\usage{
hclust.to.tree(hclust)
}
\arguments{
\item{hclust}{object of the hclust package}
}
\value{
parition matrix
}
\description{
Computes the partition matrix of an object in the hclust package.
}
