% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_data.R
\name{nested_data}
\alias{nested_data}
\title{Nested data simulation}
\usage{
nested_data(n, p)
}
\arguments{
\item{n}{number of points to be simulated}

\item{p}{size of the points to be simulated}
}
\value{
data matrix
}
\description{
Simulates a nested dataset
}
