% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNC_dissim.R
\name{NNC_dissim}
\alias{NNC_dissim}
\title{Nearest Neighbors Chain algorithm (dissimilarity)}
\usage{
NNC_dissim(D = 0, X = 0, method = "average")
}
\arguments{
\item{D}{dissimilarities matrix if already computed}

\item{X}{Data matrix (only if the similarity matrix S is not given)}

\item{method}{method for calculating dissimilarities between clusters (e.g. "single", "complete" ..), default = "average"}
}
\value{
list containing partition matrix and vector of cophenetic distances
}
\description{
Performs a ascending hierarchical classification using the Nearest-Neighbor Chain algorithm from the dissimilarity matrix.
}
\examples{
X = as.matrix(iris[-5])
method = "single"
P = NNC_dissim(X=X, method = "median")


}
