% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Similarities matrix}
\usage{
sim(X, Kernel.FUN = "linear", ...)
}
\arguments{
\item{X}{Data matrix}

\item{Kernel.FUN}{kernel function to be used if we want to calculate S from X (default :"linear")}

\item{...}{other parameters that will be passed to the Kernel procedure (see Kernel procedure for details)}
}
\value{
similarity matrix calculated from X
}
\description{
Computes the similarities matrix from a data matrix
}
